requirements.txt
# 核心依赖
python>=3.8

# LLM和API客户端
openai>=1.0.0
requests>=2.25.0

# 向量数据库和嵌入
chromadb>=0.4.0
sentence-transformers>=2.2.0
torch>=1.9.0

# 数据处理和序列化
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0

# 文本处理
nltk>=3.6.0
regex>=2022.3.0

# 配置管理
pydantic>=1.9.0
python-dotenv>=0.19.0

# 测试框架
pytest>=6.2.0

# Web框架（可选）
fastapi>=0.68.0
uvicorn>=0.15.0